feat(blog): responsive layout fixes and component consolidation
- Fix grid-cols-2 to grid-cols-1 md:grid-cols-2 in WelcomeSection and AboutMeSection for small screens
- Add min-h viewport and justify-center to WelcomeSection
- Extract PanelImage as reusable component
- Consolidate all components from src/app/components into src/components with per-component folders
