+++ apps/blog/src/app/(home)/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 import WelcomeSection from './WelcomeSection/WelcomeSection';
-import StickyHeader from '../components/StickyHeader';
+import StickyHeader from '@/components/StickyHeader/StickyHeader';
 import LandingSection from './LandingSection/LandingSection';
 import LandingSectionContextProvider from './LandingSection/LandingSection.Context';
 import AboutMeSection from './AboutMeSection/AboutMeSection';
