+++ apps/blog/src/app/blog/BlogPostCard.tsx
@@ -9,7 +9,7 @@ import {
 } from '@abbottland/fui-components';
 import { CalendarIcon, ClockIcon } from '@radix-ui/react-icons';
 import type { BlogCategory } from '../../types/blog';
-import BlogPostBannerImage from '../components/BlogPostBannerImage';
+import BlogPostBannerImage from '@/components/BlogPostBannerImage/BlogPostBannerImage';
 
 interface BlogPostCardProps {
   title: string;
