+++ apps/blog/src/app/blog/HeroBlogPostCard.tsx
@@ -8,7 +8,7 @@ import {
 } from '@abbottland/fui-components';
 import { CalendarIcon, ClockIcon } from '@radix-ui/react-icons';
 import type { BlogPost } from '../../types/blog';
-import BlogPostBannerImage from '../components/BlogPostBannerImage';
+import BlogPostBannerImage from '@/components/BlogPostBannerImage/BlogPostBannerImage';
 
 interface HeroBlogPostCardProps {
   post: BlogPost;
