+++ apps/blog/src/components/GradientMockImage/GradientMockImage.tsx
